airflow/requirements.txt
elasticsearch==8.13.0
apache-airflow==2.9.0
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-elasticsearch[common.sql]==5.3.4
langdetect==1.0.9
deep_translator==1.11.4
nltk==3.8.1
spacy==3.7.4
ijson==3.3.0
exif==1.6.0
pillow==11.0.0
opencv-python-headless==4.10.0.84
opencv-contrib-python-headless==4.10.0.84
transformers==4.46.2
torch==2.5.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1.tar.gz